Include the default post voting component in EAEmojisVoteOnPost
